Client package for v2 of the SmartFile API
Add 'api' which is the Python package that interfaces with v2 of the
SmartFile API.  Also, add unit tests for the package.  The unit tests
can be run as:  python -m unittest tests

This commit only provides a subset of the SmartFile API:
- SmartFileUser (create, list, update, delete)
- SmartFilePath (upload, delete, list (alias to path_tree.list)
- SmartFilePathOperApi (remove)
- SmartFilePathTreeApi (list)

Instantiating SmartFileAPI provides access to all of the above API
segments.  For example, api = SmartFileAPI() gives:
- SmartFileUser:  api.user
- SmartFilePath:  api.path
- SmartFilePathOperApi:  api.path_oper
- SmartFilePathTreeApi:  api.path_tree