Change how key and password are acquired
If the key and password are not provided when instantiating the API
class or any specific endpoint class, then attempt to get them from the
environment.  The key and password in the environment are
SMARTFILE_API_KEY SMARTFILE_API_PASS, respectively.  Also, remove the
attempt in the unit tests to look for API_KEY and API_PASS as they are
no longer needed.