Generate exceptions upon response error
Wrap the base CRUD calls with a decorator to raise either
SmartFileConnException if Requests raises an exception or
SmartFileResponseException if the HTTP status code is 400 or greater.
Both exceptions sub-class SmartFileException, which is a sub-class of
Exception, to allow a program to easily catch any SmartFile-based
exception.

Add a unit test that generates an exception by attempting to list a
non-existent path.