Clean up after unit test
Delete the file uploaded in PathTreeTestCase.test_file_list() to prevent
test_file_non_existent_list() from failing as the file was not
non-existent.

Delete the file uploaded in PathTestCase.test_file_download()
immediately after the test of the status code being 200.  This should
prevent a leftover if the files did not compare as the same.  It is best
to clean as soon as possible in case a unit test fails and leaves
leftovers.