Handle throttled responses from the server
When the client hits the limit of number of API requests to the server,
the server throttle the client.  It does this by responding with a
status code of 503 and sets a header (x-throttle) with the time
remaining before the throttle is removed.  To handle this, the request
that receives a 503 will wait until the throttle is removed by making
subsequent attempts with a sleep in between of the time found within the
header.  Due to probable differences in the system clocks between the
server and the client, the first request after the 503 may not succeed.

If the caller of the API does not wish to have the client code handle
the 503, then instantiating an API endpoint with throttle_wait=False
will prevent any attempt to wait.

Remove the sleep between each unit test as the throttle handling takes
care of this now.