Prevent running unit tests against production
Unless SMARTFILE_RUN_TESTS is set within the environment, the unit tests
will not run.  This is to prevent accidentally running them against the
production server by users of this code base.  Although the code
attempts to not leave leftovers nor make changes that detrimentally
affect a live site, it is best to prevent it from being run without the
user considering it a bit.